docs(nuxt): make nuxt example more idiomatic
