+++ examples/nuxt/.env.example
@@ -1,0 +1,2 @@
+GITHUB_CLIENT_ID=
+GITHUB_CLIENT_SECRET=
