+++ examples/nuxt/server/api/user.get.ts
@@ -1,0 +1,5 @@
+export default defineEventHandler(async (event) => {
+    const authRequest = auth.handleRequest(event);
+    const { user } = await authRequest.validateUser();
+    return { user };
+});
